+++ spec/concurrent/configuration_spec.rb
@@ -26,24 +26,45 @@ module Concurrent
         Concurrent::Delay.new(executor: :immediate){ Concurrent::TimerSet.new })
     end
 
-    it 'creates a global timer pool' do
-      expect(Concurrent.global_timer_set).not_to be_nil
-      expect(Concurrent.global_timer_set).to respond_to(:post)
-    end
+    context 'global executors' do
 
-    context 'global fast executor' do
+      it 'creates a global timer set' do
+        expect(Concurrent.global_timer_set).not_to be_nil
+        expect(Concurrent.global_timer_set).to respond_to(:post)
+      end
 
-      specify 'reader creates a default pool when first called if none exists' do
+      it 'creates a global fast executor' do
         expect(Concurrent.global_fast_executor).not_to be_nil
         expect(Concurrent.global_fast_executor).to respond_to(:post)
       end
-    end
 
-    context 'global io executor' do
-
-      specify 'reader creates a default pool when first called if none exists' do
+      it 'creates a global io executor' do
         expect(Concurrent.global_io_executor).not_to be_nil
         expect(Concurrent.global_io_executor).to respond_to(:post)
+      end
+
+      specify '#shutdown_global_executors acts on all global executors' do
+        expect(Concurrent.global_fast_executor).to receive(:shutdown).with(no_args)
+        expect(Concurrent.global_io_executor).to receive(:shutdown).with(no_args)
+        expect(Concurrent.global_timer_set).to receive(:shutdown).with(no_args)
+        Concurrent.shutdown_global_executors
+      end
+
+      specify '#kill_global_executors acts on all global executors' do
+        expect(Concurrent.global_fast_executor).to receive(:kill).with(no_args)
+        expect(Concurrent.global_io_executor).to receive(:kill).with(no_args)
+        expect(Concurrent.global_timer_set).to receive(:kill).with(no_args)
+        Concurrent.kill_global_executors
+      end
+
+      context '#wait_for_global_executors_termination' do
+
+        it 'acts on all global executors' do
+          expect(Concurrent.global_fast_executor).to receive(:wait_for_termination).with(0.1)
+          expect(Concurrent.global_io_executor).to receive(:wait_for_termination).with(0.1)
+          expect(Concurrent.global_timer_set).to receive(:wait_for_termination).with(0.1)
+          Concurrent.wait_for_global_executors_termination(0.1)
+        end
       end
     end
   end