+++ CHANGELOG.md
@@ -21,7 +21,7 @@
 * Simpler `:executor` option syntax for all abstractions which support this option
 * Added `Executor#auto_terminate?` predicate method (for thread pools)
 * Added `at_exit` handler to `TimerSet`
-* Simplified auto-termination of the global thread pools
+* Simplified auto-termination of the global executors
 * Deprecated terms "task pool" and "operation pool"
   - New terms are "io executor" and "fast executor"
   - New functions added with new names
@@ -30,6 +30,7 @@
   - Old functions still exist with deprecation warnings
   - New functions have updated names as appropriate
 * All high-level abstractions default to the "io executor"
+* Added shutdown/kill/wait_for_termination variants for global executors
 
 ## Current Release v0.8.0 (25 January 2015)